refactor: CustomTabBar 공통 컴포넌트화 및 import 경로 수정 #19
